+++ src/app/pages/tabs/tabs.module.ts
@@ -7,6 +7,8 @@ import { IonicModule }                      from '@ionic/angular';
 import { AgmCoreModule }                    from '@agm/core';
 import { AgmSnazzyInfoWindowModule}         from '@agm/snazzy-info-window';
 
+import { AdsenseModule }                    from 'ng2-adsense';
+
 import { TabsPageRoutingModule }            from './tabs.router.module';
 
 import { TabsPage }                         from './tabs.page';
@@ -28,6 +30,11 @@ import { TosPageModule }                    from '@app-pages/tos/tos.module';
     TabsPageRoutingModule,
     AgmCoreModule.forRoot({ apiKey: environment.firebase.apiKey }),
     AgmSnazzyInfoWindowModule,
+    AdsenseModule.forRoot({
+      adClient: environment.adSenseConfig.google_ad_client,
+      adSlot: environment.adSenseConfig.google_ad_slot,
+      pageLevelAds: environment.adSenseConfig.enable_page_level_ads
+    }),
     HomePageModule,
     CoffeeMapPageModule,
     UserSettingsPageModule,